+++ CHANGELOG.md
@@ -7,6 +7,12 @@ The format is based on
 project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+* Changed the `red(..)`, `green(..)`, and `blue(..)` functions to
+  always return integers. Issue #114, thanks @Keats.
+
+
 ## Release 0.22.0 - 2021-06-20
 
 Progress: 3992 of 6180 tests passed in dart-sass compatibility mode.